new navbar

+++ src/app/components/navbar/navbar.component.ts
@@ -3,6 +3,7 @@ import { AuthService } from "../../services/auth.service";
 import { Subscription } from "rxjs";
 import { Router } from "@angular/router";
 
+
 @Component({
   selector: 'app-navbar',
   templateUrl: './navbar.component.html',
@@ -10,25 +11,52 @@ import { Router } from "@angular/router";
 })
 export class NavbarComponent implements OnInit, OnDestroy {
 
+  userType: number;
+  navBarActive = true;
+  blurOverlay = false;
 
-  userType
   sub: Subscription = new Subscription()
-  constructor(
-    private authService: AuthService,
-    private router: Router) {
-  }
+
+  constructor(private authService: AuthService, private  route: Router) {}
 
   ngOnInit(): void {
+    window.innerWidth >= 768 ? this.navBarActive = true : this.navBarActive = false;
     this.sub.add(this.authService.user$.subscribe(type => this.userType = type));
+  }
+
+  navigate(navbar, ham?, url?): void {
+    if (navbar?.classList.contains('nav-bar-links-active')) {
+      navbar.classList.remove('nav-bar-links-active');
+      this.navBarActive = !this.navBarActive;
+      this.blurOverlay = !this.blurOverlay;
+    }
+    if (ham) {
+      this.navBarActive ? ham.style.transform = 'rotate(180deg)' : ham.style.transform = 'rotate(0deg)';
+    }
+    if (url) {
+      this.route.navigate(url).then();
+    }
+  }
+
+  activeNav(nav, ham): void {
+    nav.classList.toggle('nav-bar-links-active');
+    this.blurOverlay = !this.blurOverlay;
+    this.navBarActive = !this.navBarActive;
+    this.navBarActive ? ham.style.transform = 'rotate(180deg)' : ham.style.transform = 'rotate(0deg)';
+
+  }
+
+  logout(navbar) {
+    let answer = confirm('Are you sure you want to log-out?');
+    if (answer) {
+      navbar.classList.toggle('nav-bar-links-active');
+      sessionStorage.clear();
+      location.href = '/home'
+    }
   }
 
   ngOnDestroy(): void {
     this.sub.unsubscribe();
   }
 
-
-  logout() {
-    sessionStorage.clear();
-    location.href = '/home'
-  }
 }
